# contracts/product-api/v2/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResearchArtifactSummaryDto": {
        "additionalProperties": false,
        "properties": {
          "artifact_content_fingerprint": {
            "title": "Artifact Content Fingerprint",
            "type": "string"
          },
          "artifact_profile": {
            "title": "Artifact Profile",
            "type": "string"
          },
          "artifact_schema_version": {
            "title": "Artifact Schema Version",
            "type": "integer"
          },
          "candidate_count": {
            "title": "Candidate Count",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "dataset_snapshot_fingerprint": {
            "title": "Dataset Snapshot Fingerprint",
            "type": "string"
          },
          "instrument_ids": {
            "items": {
              "type": "string"
            },
            "title": "Instrument Ids",
            "type": "array"
          },
          "market_row_count": {
            "title": "Market Row Count",
            "type": "integer"
          },
          "published_series_count": {
            "title": "Published Series Count",
            "type": "integer"
          },
          "research_result_content_fingerprint": {
            "title": "Research Result Content Fingerprint",
            "type": "string"
          },
          "research_result_fingerprint": {
            "title": "Research Result Fingerprint",
            "type": "string",
            "x-onlyalpha-reference-kind": "RESEARCH_RESULT",
            "x-onlyalpha-reference-locator-kind": "SHA256",
            "x-onlyalpha-reference-schema-version": 1
          },
          "research_result_plan_fingerprint": {
            "title": "Research Result Plan Fingerprint",
            "type": "string"
          },
          "research_result_schema_version": {
            "title": "Research Result Schema Version",
            "type": "integer"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "schema_version": {
            "const": 2,
            "default": 2,
            "title": "Schema Version",
            "type": "integer"
          },
          "signal_series_count": {
            "title": "Signal Series Count",
            "type": "integer"
          },
          "statistics_count": {
            "title": "Statistics Count",
            "type": "integer"
          }
        },
        "required": [
          "research_result_plan_fingerprint",
          "research_result_content_fingerprint",
          "research_result_fingerprint",
          "dataset_snapshot_fingerprint",
          "artifact_content_fingerprint",
          "research_result_schema_version",
          "artifact_profile",
          "artifact_schema_version",
          "statistics_count",
          "row_count",
          "candidate_count",
          "published_series_count",
          "signal_series_count",
          "market_row_count",
          "instrument_ids",
          "created_at"
        ],
        "title": "ResearchArtifactSummaryDto",
        "type": "object"
      },
...
      "ResearchErrorDto": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "schema_version": {
            "const": 2,
            "default": 2,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ResearchErrorDto",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OnlyAlpha Product API",
    "version": "2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/research/artifacts/{research_result_fingerprint}": {
      "get": {
        "operationId": "artifact_summary_api_v2_research_artifacts__research_result_fingerprint__get",
        "parameters": [
          {
            "in": "path",
            "name": "research_result_fingerprint",
            "required": true,
            "schema": {
              "title": "Research Result Fingerprint",
              "type": "string",
              "x-onlyalpha-reference-kind": "RESEARCH_RESULT",
              "x-onlyalpha-reference-locator-kind": "SHA256",
              "x-onlyalpha-reference-schema-version": 1
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchArtifactSummaryDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Invalid Research query"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Exact Artifact or Statistics identity not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Scientific evidence is unavailable for this profile"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Research Artifact verification failed"
          }
        },
        "summary": "Artifact Summary",
        "tags": [
          "research-artifacts"
        ],
        "x-onlyalpha-agent-operation": {
          "identity_requirements": [
            "research_result_fingerprint"
          ],
          "owning_authority_references": [
            {
              "locator_kind": "SHA256",
              "reference_kind": "RESEARCH_RESULT",
              "reference_schema_version": 1,
              "response_field": "research_result_fingerprint"
            }
          ],
          "recovery_class": "IMMUTABLE_EXACT_QUERY",
          "requires_product_command_id": false,
          "schema_version": 1,
          "tool_class": "RESEARCH_EVIDENCE_QUERY"
        }
      }
    },
    "/api/v2/research/artifacts/{research_result_fingerprint}/candidates": {
      "get": {
        "operationId": "candidate_catalog_api_v2_research_artifacts__research_result_fingerprint__candidates_get",
        "parameters": [
          {
            "in": "path",
            "name": "research_result_fingerprint",
            "required": true,
            "schema": {
              "title": "Research Result Fingerprint",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchCandidateCatalogDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Invalid Research query"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Exact Artifact or Statistics identity not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Scientific evidence is unavailable for this profile"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Research Artifact verification failed"
          }
        },
        "summary": "Candidate Catalog",
        "tags": [
          "research-artifacts"
        ]
      }
    },
    "/api/v2/research/artifacts/{research_result_fingerprint}/candidates/{candidate_fingerprint}/graph": {
      "get": {
        "operationId": "candidate_graph_api_v2_research_artifacts__research_result_fingerprint__candidates__candidate_fingerprint__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "research_result_fingerprint",
            "required": true,
            "schema": {
              "title": "Research Result Fingerprint",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_fingerprint",
            "required": true,
            "schema": {
              "title": "Candidate Fingerprint",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchCandidateGraphDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Invalid Research query"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Exact Artifact or Statistics identity not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Scientific evidence is unavailable for this profile"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Research Artifact verification failed"
          }
        },
        "summary": "Candidate Graph",
        "tags": [
          "research-artifacts"
        ]
      }
    },
    "/api/v2/research/artifacts/{research_result_fingerprint}/market/series": {
      "get": {
        "operationId": "market_series_api_v2_research_artifacts__research_result_fingerprint__market_series_get",
        "parameters": [
          {
            "in": "path",
            "name": "research_result_fingerprint",
            "required": true,
            "schema": {
              "title": "Research Result Fingerprint",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "title": "Instrument Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from_ts_event_ns",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(?:0|-?[1-9][0-9]*)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Ts Event Ns"
            }
          },
          {
            "in": "query",
            "name": "to_ts_event_ns",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(?:0|-?[1-9][0-9]*)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Ts Event Ns"
            }
          },
          {
            "in": "query",
            "name": "after_ts_event_ns",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(?:0|-?[1-9][0-9]*)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Ts Event Ns"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchScientificSeriesPageDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Invalid Research query"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Exact Artifact or Statistics identity not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Scientific evidence is unavailable for this profile"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Research Artifact verification failed"
          }
        },
        "summary": "Market Series",
        "tags": [
          "research-artifacts"
        ]
      }
    },
    "/api/v2/research/artifacts/{research_result_fingerprint}/signals/{candidate_fingerprint}/{role}/series": {
      "get": {
        "operationId": "signal_series_api_v2_research_artifacts__research_result_fingerprint__signals__candidate_fingerprint___role__series_get",
        "parameters": [
          {
            "in": "path",
            "name": "research_result_fingerprint",
            "required": true,
            "schema": {
              "title": "Research Result Fingerprint",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_fingerprint",
            "required": true,
            "schema": {
              "title": "Candidate Fingerprint",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "role",
            "required": true,
            "schema": {
              "title": "Role",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "title": "Instrument Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from_ts_event_ns",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(?:0|-?[1-9][0-9]*)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Ts Event Ns"
            }
          },
          {
            "in": "query",
            "name": "to_ts_event_ns",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(?:0|-?[1-9][0-9]*)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Ts Event Ns"
            }
          },
          {
            "in": "query",
            "name": "after_ts_event_ns",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(?:0|-?[1-9][0-9]*)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Ts Event Ns"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchScientificSeriesPageDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Invalid Research query"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Exact Artifact or Statistics identity not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Scientific evidence is unavailable for this profile"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Research Artifact verification failed"
          }
        },
        "summary": "Signal Series",
        "tags": [
          "research-artifacts"
        ]
      }
    },
    "/api/v2/research/artifacts/{research_result_fingerprint}/statistics": {
      "get": {
        "operationId": "statistics_catalog_api_v2_research_artifacts__research_result_fingerprint__statistics_get",
        "parameters": [
          {
            "in": "path",
            "name": "research_result_fingerprint",
            "required": true,
            "schema": {
              "title": "Research Result Fingerprint",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchStatisticsCatalogDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Invalid Research query"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Exact Artifact or Statistics identity not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Scientific evidence is unavailable for this profile"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Research Artifact verification failed"
          }
        },
        "summary": "Statistics Catalog",
        "tags": [
          "research-artifacts"
        ]
      }
    },
...
    "/api/v2/research/artifacts/{research_result_fingerprint}/variables": {
      "get": {
        "operationId": "variable_catalog_api_v2_research_artifacts__research_result_fingerprint__variables_get",
        "parameters": [
          {
            "in": "path",
            "name": "research_result_fingerprint",
            "required": true,
            "schema": {
              "title": "Research Result Fingerprint",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchPublishedSeriesCatalogDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Invalid Research query"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Exact Artifact or Statistics identity not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Scientific evidence is unavailable for this profile"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchErrorDto"
                }
              }
            },
            "description": "Research Artifact verification failed"
          }
        },
        "summary": "Variable Catalog",
        "tags": [
          "research-artifacts"
        ]
      }
    },
...
  }
}